requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.1
mlflow==3.9.0
skops==0.11.0
dvc==3.40.0
pyyaml==6.0.1
